cleanup crates && add deprecation notices
