+++ Cargo.toml
@@ -41,7 +41,6 @@ members = [
     "crates/claim",
     "crates/miner",
     "crates/state",
-    "crates/txn",
     "crates/wallet",
     "crates/events",
     "crates/ev_mem_pool",
@@ -56,7 +55,6 @@ members = [
     "crates/node",
     "crates/storage",
     "crates/syncing",
-    "crates/verifiable",
     "crates/ledger",
     "crates/accountable",
     "crates/ownable",
