+++ crates/block/src/block.rs
@@ -1,38 +1,50 @@
 // This file contains code for creating blocks to be proposed, including the
 // genesis block and blocks being mined.
-#![allow(unused_imports)]
-#![allow(dead_code)]
-use std::fmt;
 
-use accountable::accountable::Accountable;
-use claim::claim::Claim;
-use log::info;
-use primitives::types::{Epoch, RawSignature, GENESIS_EPOCH};
-use rand::Rng;
-use reward::reward::{Reward, GENESIS_REWARD, NUMBER_OF_BLOCKS_PER_EPOCH};
+use primitives::types::{Epoch, RawSignature, GENESIS_EPOCH, SECOND, VALIDATOR_THRESHOLD};
+use reward::reward::{Reward, NUMBER_OF_BLOCKS_PER_EPOCH};
 use ritelinked::LinkedHashMap;
 use serde::{Deserialize, Serialize};
 use sha256::digest;
 use state::state::NetworkState;
-use txn::txn::Txn;
-use verifiable::verifiable::Verifiable;
+use std::fmt;
+use vrrb_core::accountable::Accountable;
+use vrrb_core::claim::Claim;
+use vrrb_core::txn::Txn;
+use vrrb_core::verifiable::Verifiable;
+
+#[cfg(mainnet)]
+use crate::genesis;
+
+#[cfg(mainnet)]
+use reward::reward::GENESIS_REWARD;
 
 use crate::{
-    genesis,
     header::BlockHeader,
     invalid::{InvalidBlockError, InvalidBlockErrorReason},
 };
 
-pub const NANO: u128 = 1;
-pub const MICRO: u128 = NANO * 1000;
-pub const MILLI: u128 = MICRO * 1000;
-pub const SECOND: u128 = MILLI * 1000;
-const VALIDATOR_THRESHOLD: f64 = 0.60;
 pub const GROSS_UTILITY_PERCENTAGE: f64 = 0.01;
 pub const PERCENTAGE_CHANGE_SUPPLY_CAP: f64 = 0.25;
+
 pub type CurrentUtility = i128;
 pub type NextEpochAdjustment = i128;
 
+pub struct MineArgs<'a> {
+    /// The claim entitling the miner to mine the block.
+    pub claim: Claim,
+    pub last_block: Block,
+    /// The last block, which contains the current block reward.
+    pub txns: LinkedHashMap<String, Txn>,
+    pub claims: LinkedHashMap<String, Claim>,
+    pub claim_map_hash: Option<String>,
+    pub reward: &'a mut Reward,
+    pub network_state: &'a NetworkState,
+    pub neighbors: Option<Vec<BlockHeader>>,
+    pub abandoned_claim: Option<Claim>,
+    pub signature: String,
+    pub epoch: Epoch,
+}
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 #[repr(C)]
@@ -63,7 +75,6 @@ pub struct Block {
     /// Adjustment For Next Epoch
     pub adjustment_for_next_epoch: Option<NextEpochAdjustment>,
 }
-
 
 impl Block {
     // Returns a result with either a tuple containing the genesis block and the
@@ -121,20 +132,32 @@ impl Block {
     /// The mine method is used to generate a new block (and an updated account
     /// state with the reward set to the miner wallet's balance), this will
     /// also update the network state with a new confirmed state.
-    #[allow(clippy::too_many_arguments)]
     pub fn mine(
-        claim: Claim,      // The claim entitling the miner to mine the block.
-        last_block: Block, // The last block, which contains the current block reward.
-        txns: LinkedHashMap<String, Txn>,
-        claims: LinkedHashMap<String, Claim>,
-        claim_map_hash: Option<String>,
-        reward: &mut Reward,
-        network_state: &NetworkState,
-        neighbors: Option<Vec<BlockHeader>>,
-        abandoned_claim: Option<Claim>,
-        signature: String,
-        epoch: Epoch,
+        args: MineArgs,
+        // claim: Claim,      // The claim entitling the miner to mine the block.
+        // last_block: Block, // The last block, which contains the current block reward.
+        // txns: LinkedHashMap<String, Txn>,
+        // claims: LinkedHashMap<String, Claim>,
+        // claim_map_hash: Option<String>,
+        // reward: &mut Reward,
+        // network_state: &NetworkState,
+        // neighbors: Option<Vec<BlockHeader>>,
+        // abandoned_claim: Option<Claim>,
+        // signature: String,
+        // epoch: Epoch,
     ) -> (Option<Block>, NextEpochAdjustment) {
+        let claim = args.claim;
+        let last_block = args.last_block;
+        let txns = args.txns;
+        let claims = args.claims;
+        let claim_map_hash = args.claim_map_hash;
+        let reward = args.reward;
+        let network_state = args.network_state;
+        let neighbors = args.neighbors;
+        let abandoned_claim = args.abandoned_claim;
+        let signature = args.signature;
+        let epoch = args.epoch;
+
         // TODO: Replace with Tx Trie Root
         let txn_hash = {
             let mut txn_vec = vec![];
@@ -167,7 +190,6 @@ impl Block {
         } else {
             block_utility = utility_amount + last_block.utility;
         }
-
 
         // TODO: Fix after replacing neighbors and tx hash/claim hash with respective
         // Trie Roots
@@ -303,34 +325,29 @@ impl Verifiable for Block {
         true
     }
 
-    #[allow(unused_variables)]
     fn valid(
         &self,
         item: &Self::Item,
         dependencies: &Self::Dependencies,
     ) -> Result<bool, Self::Error> {
         if self.header.block_height > item.header.block_height + 1 {
-            return Err(Self::Error {
-                details: InvalidBlockErrorReason::BlockOutOfSequence,
-            });
+            return Err(Self::Error::new(
+                InvalidBlockErrorReason::BlockOutOfSequence,
+            ));
         }
 
         if self.header.block_height <= item.header.block_height {
-            return Err(Self::Error {
-                details: InvalidBlockErrorReason::NotTallestChain,
-            });
+            return Err(Self::Error::new(InvalidBlockErrorReason::NotTallestChain));
         }
 
         if self.header.block_nonce != item.header.next_block_nonce {
-            return Err(Self::Error {
-                details: InvalidBlockErrorReason::InvalidBlockNonce,
-            });
+            return Err(Self::Error::new(InvalidBlockErrorReason::InvalidBlockNonce));
         }
 
         if self.header.block_reward.get_amount() != item.header.next_block_reward.get_amount() {
-            return Err(Self::Error {
-                details: InvalidBlockErrorReason::InvalidBlockReward,
-            });
+            return Err(Self::Error::new(
+                InvalidBlockErrorReason::InvalidBlockReward,
+            ));
         }
 
         if let Some((hash, pointers)) =
@@ -343,32 +360,28 @@ impl Verifiable for Block {
                     .get_pointer(self.header.block_nonce as u128)
                 {
                     if pointers != claim_pointer {
-                        return Err(Self::Error {
-                            details: InvalidBlockErrorReason::InvalidClaimPointers,
-                        });
+                        return Err(Self::Error::new(
+                            InvalidBlockErrorReason::InvalidClaimPointers,
+                        ));
                     }
                 } else {
-                    return Err(Self::Error {
-                        details: InvalidBlockErrorReason::InvalidClaimPointers,
-                    });
+                    return Err(Self::Error::new(
+                        InvalidBlockErrorReason::InvalidClaimPointers,
+                    ));
                 }
             } else {
-                return Err(Self::Error {
-                    details: InvalidBlockErrorReason::InvalidClaimPointers,
-                });
+                return Err(Self::Error::new(
+                    InvalidBlockErrorReason::InvalidClaimPointers,
+                ));
             }
         }
 
         if self.header.last_hash != item.hash {
-            return Err(Self::Error {
-                details: InvalidBlockErrorReason::InvalidLastHash,
-            });
+            return Err(Self::Error::new(InvalidBlockErrorReason::InvalidLastHash));
         }
 
         if self.header.claim.valid(&None, &(None, None)).is_err() {
-            return Err(Self::Error {
-                details: InvalidBlockErrorReason::InvalidClaim,
-            });
+            return Err(Self::Error::new(InvalidBlockErrorReason::InvalidClaim));
         }
 
         Ok(true)
@@ -386,33 +399,27 @@ impl Verifiable for Block {
         );
 
         if self.header.block_height != 0 {
-            return Err(Self::Error {
-                details: InvalidBlockErrorReason::InvalidBlockHeight,
-            });
+            return Err(Self::Error::new(
+                InvalidBlockErrorReason::InvalidBlockHeight,
+            ));
         }
 
         if self.header.last_hash != genesis_last_hash {
-            return Err(Self::Error {
-                details: InvalidBlockErrorReason::InvalidLastHash,
-            });
+            return Err(Self::Error::new(InvalidBlockErrorReason::InvalidLastHash));
         }
 
         if self.hash != genesis_state_hash {
-            return Err(Self::Error {
-                details: InvalidBlockErrorReason::InvalidStateHash,
-            });
+            return Err(Self::Error::new(InvalidBlockErrorReason::InvalidStateHash));
         }
 
         if self.header.claim.valid(&None, &(None, None)).is_err() {
-            return Err(Self::Error {
-                details: InvalidBlockErrorReason::InvalidClaim,
-            });
+            return Err(Self::Error::new(InvalidBlockErrorReason::InvalidClaim));
         }
 
         if self.header.verify().is_err() {
-            return Err(Self::Error {
-                details: InvalidBlockErrorReason::InvalidBlockSignature,
-            });
+            return Err(Self::Error::new(
+                InvalidBlockErrorReason::InvalidBlockSignature,
+            ));
         }
 
         let mut valid_data = true;
@@ -424,9 +431,7 @@ impl Verifiable for Block {
         });
 
         if !valid_data {
-            return Err(Self::Error {
-                details: InvalidBlockErrorReason::InvalidTxns,
-            });
+            return Err(Self::Error::new(InvalidBlockErrorReason::InvalidTxns));
         }
 
         Ok(true)
