+++ crates/claim/Cargo.toml
@@ -8,7 +8,5 @@ edition = "2021"
 [dependencies]
 serde_json = "1.0.64"
 serde = { version = "1.0.101", features = ["derive"] }
-verifiable = { path = "../verifiable" }
 sha256 = "1.0.2"
-ownable = { path = "../ownable" }
-noncing = { path = "../noncing" }
+vrrb_core = { path = "../vrrb_core" }
