+++ crates/ledger/Cargo.toml
@@ -11,8 +11,4 @@ sha256 = "1.0.2"
 serde_json = "1.0.64"
 hex = "*"
 serde = { version = "1.0.101", features = ["derive"] }
-verifiable = { path = "../verifiable" }
-claim = { path = "../claim" }
-ownable = { path = "../ownable" }
-noncing = { path = "../noncing" }
-
+vrrb_core = { path = "../vrrb_core" }
