+++ crates/mempool/Cargo.toml
@@ -9,7 +9,7 @@ edition = "2021"
 evmap = "10.0.2"
 serde_json = "1.0.64"
 serde = { version = "1.0.101", features = ["derive"] }
-txn = { path = "../txn" }
+vrrb_core = { path = "../vrrb_core" }
 state = { path = "../state" }
 left-right = "0.11.5"
 indexmap = "1.9.1"
