+++ crates/miner/src/miner.rs
@@ -12,8 +12,6 @@ use std::{
 /// checkpoints in the state.
 use block::block::Block;
 use block::header::BlockHeader;
-use claim::claim::Claim;
-use noncing::nonceable::Nonceable;
 use pool::pool::{Pool, PoolKind};
 use primitives::types::Epoch;
 use reward::reward::Reward;
@@ -21,7 +19,10 @@ use ritelinked::LinkedHashMap;
 use serde::{Deserialize, Serialize};
 use sha256::digest;
 use state::state::NetworkState;
-use txn::txn::Txn;
+use vrrb_core::claim::Claim;
+use vrrb_core::nonceable::Nonceable;
+use vrrb_core::txn::Txn;
+
 pub const VALIDATOR_THRESHOLD: f64 = 0.60;
 pub const NANO: u128 = 1;
 pub const MICRO: u128 = NANO * 1000;
