+++ crates/node/Cargo.toml
@@ -20,7 +20,6 @@ miner = { path = "../miner" }
 claim = { path = "../claim" }
 ledger = { path = "../ledger" }
 wallet = { path = "../wallet" }
-txn = { path = "../txn" }
 state = { path = "../state" }
 validator = { path = "../validator" }
 events = { path = "../events" }
