+++ crates/noncing/src/lib.rs
@@ -1,3 +1,4 @@
+#[deprecated(note = "use the trait defined at crates/vrrb_core instead")]
 pub mod nonceable;
 
 #[cfg(test)]
