+++ crates/reward/Cargo.toml
@@ -6,10 +6,11 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-rand = {version = "0.6", features = ["std"]}
+rand = { version = "0.6", features = ["std"] }
 strum = "0.21.0"
 strum_macros = "0.21.0"
 serde_json = "1.0.64"
 serde = { version = "1.0.101", features = ["derive"] }
-accountable = { path = "../accountable" }
 primitives = { path = "../primitives" }
+vrrb_core = { path = "../vrrb_core" }
+
