+++ crates/state/src/state.rs
@@ -3,27 +3,20 @@ use std::fs;
 
 /// This module contains the Network State struct (which will be replaced with
 /// the Left-Right State Trie)
-use accountable::accountable::Accountable;
+use vrrb_core::accountable::Accountable;
 use claim::claim::Claim;
 use ledger::ledger::Ledger;
 use log::info;
-use noncing::nonceable::Nonceable;
-use ownable::ownable::Ownable;
 use reward::reward::Reward;
 use ritelinked::LinkedHashMap;
 use serde::{Deserialize, Serialize};
 use sha256::digest;
+use vrrb_core::nonceable::Nonceable;
+use vrrb_core::ownable::Ownable;
 
 use crate::types::{
-    CreditsHash,
-    CreditsRoot,
-    DebitsHash,
-    DebitsRoot,
-    LedgerBytes,
-    StateHash,
-    StatePath,
-    StateRewardState,
-    StateRoot,
+    CreditsHash, CreditsRoot, DebitsHash, DebitsRoot, LedgerBytes, StateHash, StatePath,
+    StateRewardState, StateRoot,
 };
 
 /// The Network State struct, contains basic information required to determine
