--- crates/txn/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "txn"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-secp256k1 = {version = "0.20.2", features = ["rand"]}
-sha256 = "1.0.2"
-serde_json = "1.0.64"
-serde = { version = "1.0.101", features = ["derive"] }
-bytebuffer = "0.2.1"
-uuid = { version = "0.8.0", features = ["serde", "v4"] }
-accountable = { path = "../accountable" }
-pool = { path = "../pool" }
-verifiable = { path = "../verifiable" }
-state = { path = "../state" }
