--- crates/txn/src/lib.rs
@@ -1,9 +1,0 @@
-pub mod txn;
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
-}
