+++ crates/validator/Cargo.toml
@@ -8,10 +8,8 @@ edition = "2021"
 [dependencies]
 serde_json = "1.0.64"
 serde = { version = "1.0.101", features = ["derive"] }
-txn = { path = "../txn" }
 pool = { path = "../pool" }
 state = { path = "../state" }
-verifiable = { path = "../verifiable" }
 mempool = { path = "../mempool" }
 left-right = "0.11.5"
 lr_trie = { path = "../lr_trie" }
@@ -22,8 +20,8 @@ bytebuffer = "0.2.1"
 blake3 = "0.3.8"
 bincode = "1.3.3"
 lrdb = { path = "../lrdb" }
+vrrb_core = { path = "../vrrb_core" }
 
 [dev-dependencies]
-validator = { path = "../validator" }
 rand = "0.8.5"
 mempool = { path = "../mempool" }
