+++ crates/validator/src/txn_validator.rs
@@ -11,10 +11,10 @@ use lrdb::Account;
 use patriecia::{db::Database, error::TrieError, inner::InnerTrie};
 #[allow(deprecated)]
 use secp256k1::{
-    Signature,
-    {Message, PublicKey, Secp256k1},
+    Signature, {Message, PublicKey, Secp256k1},
 };
-use txn::txn::Txn;
+
+use vrrb_core::txn::Txn;
 
 type Result<T> = StdResult<T, TxnValidatorError>;
