--- crates/verifiable/src/lib.rs
@@ -1,9 +1,0 @@
-pub mod verifiable;
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
-}
