+++ crates/vrrb_core/src/accountable.rs
@@ -1,4 +1,3 @@
-// FEATURE TAG(S): Tx Validation, Tx Writes to Confirm
 /// A trait that enables the type it is implemented on to provide some basic
 /// accounting functionality.
 pub trait Accountable {
