+++ crates/vrrb_core/src/lib.rs
@@ -1,5 +1,11 @@
 pub mod account;
 pub mod accountable;
+pub mod claim;
 pub mod event_router;
+pub mod ledger;
+pub mod nonceable;
+pub mod ownable;
 pub mod result;
+pub mod txn;
+pub mod verifiable;
 pub use result::*;
